LEA (`PeerConfigurator`): Split up the `types` into multiple files.
