+++ opendut-lea/src/peers/configurator/tabs/executor/mod.rs
@@ -3,7 +3,8 @@ use opendut_model::peer::executor::container::Engine;
 use opendut_model::peer::executor::ExecutorId;
 use crate::components::UserInputValue;
 use crate::peers::configurator::tabs::executor::executor_panel::ExecutorPanel;
-use crate::peers::configurator::types::{EMPTY_CONTAINER_IMAGE_ERROR_MESSAGE, UserPeerConfiguration, UserPeerExecutor, UserPeerExecutorKind};
+use crate::peers::configurator::types::executor::{UserPeerExecutor, UserPeerExecutorKind};
+use crate::peers::configurator::types::{UserPeerConfiguration, EMPTY_CONTAINER_IMAGE_ERROR_MESSAGE};
 
 mod executor_panel;
